+++ R/covm.R
@@ -8,7 +8,7 @@ covm <- function(pars,isotropic=FALSE,axes=c("x","y"))
 {
   if(is.null(pars)) { return(NULL) }
 
-  if(class(pars)=="covm") { pars <- pars@par }
+  if(class(pars)[1]=="covm") { pars <- pars@par }
 
   if(length(axes)==1)
   {
@@ -29,7 +29,7 @@ covm <- function(pars,isotropic=FALSE,axes=c("x","y"))
     else if(length(pars)==4)
     {
       sigma <- pars
-      if(class(pars)=="covm") { pars <- attr(pars,'par') }
+      if(class(pars)[1]=="covm") { pars <- attr(pars,'par') }
       else { pars <- sigma.destruct(sigma,isotropic=isotropic) }
     }
 
@@ -370,9 +370,10 @@ COV.covm <- function(sigma,n,k=1,REML=TRUE)
     # gradient matrix d sigma / d par
     grad <- J.sigma.par(par)
     # gradient matrix d par / d sigma via inverse function theorem
-    grad <- solve(grad)
+    grad <- PDsolve(grad)
 
     COV <- (grad) %*% COV %*% t(grad)
+    COV <- nant(COV,0) # 0/0 for inactive
 
     COV <- He(COV) # asymmetric errors
     dimnames(COV) <- list(names(par),names(par))