+++ R/generic.R
@@ -5,57 +5,11 @@ match.arg <- function(arg,choices,...)
   else { return(base::match.arg(arg,choices,...)) }
 }
 
-# is a package installed?
-is.installed <- function(pkg) is.element(pkg, utils::installed.packages()[,1])
-
 # does this thing exist and, if so, is it true
 true <- function(x) { !is.null(x) & !is.na(x) & x }
 
 # not in #
 "%nin%" <- function(x, table) { match(x, table, nomatch = 0) == 0 }
-
-# generic FFT functions
-FFT <- function(X,inverse=FALSE)
-{
-  if(is.null(ncol(X)) || is.na(ncol(X)))
-  {
-    if(!inverse) { X <- stats::fft(X) }
-    else { X <- stats::fft(X,inverse=TRUE)/length(X) }
-  }
-  else
-  {
-    if(!inverse) { X <- stats::mvfft(X) }
-    else { X <- stats::mvfft(X,inverse=TRUE)/nrow(X) }
-  }
-
-  return(X)
-}
-
-# fastest FFT functions... don't use on integers
-FFTW <- function(X,inverse=FALSE)
-{
-  if(is.null(ncol(X)) || is.na(ncol(X)))
-  {
-    if(!inverse) { X <- fftw::FFT(X) }
-    else { X <- fftw::IFFT(X) }
-  }
-  else
-  {
-    if(!inverse) { X <- sapply(1:ncol(X),function(j){ fftw::FFT(X[,j]) }) }
-    else { X <- sapply(1:ncol(X),function(j){ fftw::IFFT(X[,j]) }) }
-  }
-
-  return(X)
-}
-
-.onLoad <- function(...)
-{
-  # choose FFTW if installed
-  if(is.installed("fftw")) { utils::assignInMyNamespace("FFT", FFTW) }
-}
-.onAttach <- .onLoad
-
-IFFT <- function(X,plan=NULL) { FFT(X,inverse=TRUE) }
 
 composite <- function(n) { 2^ceiling(log(n,2)) }
 
@@ -91,7 +45,7 @@ mpsigamma <- function(x,deriv=0,dim=1)
 # forwarding function for list of a particular datatype
 zoom.list <- function(x,...)
 {
-  CLASS <- class(x[[1]])
+  CLASS <- class(x[[1]])[1]
   #utils::getS3method("zoom",CLASS)(x,...)
   methods::getMethod("zoom",signature=CLASS)(x,...)
 }
@@ -101,7 +55,7 @@ methods::setMethod("zoom",signature(x="list"), function(x,...) zoom.list(x,...))
 # forwarding function for list of a particular datatype
 mean.list <- function(x,...)
 {
-  CLASS <- class(x[[1]])
+  CLASS <- class(x[[1]])[1]
   utils::getS3method("mean",CLASS)(x,...)
 }
 #methods::setMethod("mean",signature(x="list"), function(x,...) mean.list(x,...))
@@ -110,7 +64,7 @@ mean.list <- function(x,...)
 # forwarding function for list of a particular datatype
 median.list <- function(x,na.rm=FALSE,...)
 {
-  CLASS <- class(x[[1]])
+  CLASS <- class(x[[1]])[1]
   utils::getS3method("median",CLASS)(x,na.rm=na.rm,...)
 }
 
@@ -118,7 +72,7 @@ median.list <- function(x,na.rm=FALSE,...)
 # forwarding function for list of a particular datatype
 plot.list <- function(x,...)
 {
-  CLASS <- class(x[[1]])
+  CLASS <- class(x[[1]])[1]
   utils::getS3method("plot",CLASS)(x,...)
 }
 #methods::setMethod("plot",signature(x="list"), function(x,...) plot.list(x,...))
@@ -126,14 +80,14 @@ plot.list <- function(x,...)
 # forwarding function for list of a particular datatype
 summary.list <- function(object,...)
 {
-  CLASS <- class(object[[1]])
+  CLASS <- class(object[[1]])[1]
   utils::getS3method("summary",CLASS)(object,...)
 }
 
 # forwarding function for list of a particular datatype
 writeShapefile.list <- function(object,folder,file=NULL,...)
 {
-  CLASS <- class(object[[1]])
+  CLASS <- class(object[[1]])[1]
   utils::getS3method("writeShapefile",CLASS)(object,folder,file=file,...)
 }
 
@@ -156,11 +110,15 @@ is.odd <- Vectorize(function(x) {x %% 2 != 0})
 # generalized covariance from -likelihood derivatives
 cov.loglike <- function(hess,grad=rep(0,nrow(hess)))
 {
+  # in case of bad derivatives, use worst-case numbers
+  grad <- nant(grad,Inf)
+  hess <- nant(hess,0)
+
   # if hessian is likely to be positive definite
   if(all(diag(hess)>0))
   {
     COV <- try(PDsolve(hess))
-    if(class(COV)=="matrix" && all(diag(COV)>0)) { return(COV) }
+    if(class(COV)[1]=="matrix" && all(diag(COV)>0)) { return(COV) }
   }
   # one of the curvatures is negative or close to negative
   # return something sensible just in case we are on a boundary and this makes sense
@@ -182,7 +140,27 @@ cov.loglike <- function(hess,grad=rep(0,nrow(hess)))
 
   EIGEN <- eigen(hess)
   values <- EIGEN$values
-  if(any(values<=0)) { warning("MLE is near a boundary or optim() failed.") }
+  if(any(values<=0))
+  {
+    # shouldn't need to warn if using ctmm.select
+    WARN <- TRUE
+    N <- sys.nframe()
+    if(N>=2)
+    {
+      for(i in 2:N)
+      {
+        CALL <- deparse(sys.call(-i))[1]
+        CALL <- grepl("ctmm.select",CALL)
+        if(CALL)
+        {
+          WARN <- FALSE
+          break
+        }
+      }
+    }
+    # warn if weren't using ctmm.select
+    if(WARN) { warning("MLE is near a boundary or optimizer failed.") }
+  }
   values <- clamp(values,0,Inf)
   vectors <- EIGEN$vectors
 
@@ -289,7 +267,7 @@ listify <- function(x)
 {
   if(is.null(x)) { return(x) }
 
-  if(class(x) != "list")
+  if(class(x)[1] != "list")
   {
     x <- list(x)
     names(x) <- attr(x[[1]],'info')$identity