+++ R/kde.R
@@ -81,9 +81,9 @@ bandwidth <- function(data,CTMM,VMM=NULL,weights=FALSE,fast=TRUE,dt=NULL,precisi
       }
     }
 
-    if(!WEIGHTS) # class(weights)=='numeric' || class(weights)=='integer' ||  # fixed weight lag information
+    if(!WEIGHTS) # class(weights)[1]=='numeric' || class(weights)[1]=='integer' ||  # fixed weight lag information
     {
-      # if(class(weights)=='numeric' || class(weights)=='integer') # fixed weights
+      # if(class(weights)[1]=='numeric' || class(weights)[1]=='integer') # fixed weights
       # { weights <- weights/sum(weights) }
       # else # uniform weights
       # { weights <- rep(1/n,n) }
@@ -402,7 +402,7 @@ akde <- function(data,CTMM,VMM=NULL,debias=TRUE,weights=FALSE,smooth=TRUE,error=
   if(length(projection(data))>1) { stop("Data not in single coordinate system.") }
   validate.grid(data,grid)
 
-  DROP <- class(data)!="list"
+  DROP <- class(data)[1] != "list"
   data <- listify(data)
   CTMM <- listify(CTMM)
   VMM <- listify(VMM)
@@ -417,7 +417,7 @@ akde <- function(data,CTMM,VMM=NULL,debias=TRUE,weights=FALSE,smooth=TRUE,error=
   for(i in 1:n)
   {
     CTMM0[[i]] <- CTMM[[i]] # original model fit
-    if(class(CTMM[[i]])=="ctmm") # calculate bandwidth etc.
+    if(class(CTMM[[i]])[1]=="ctmm") # calculate bandwidth etc.
     {
       axes <- CTMM[[i]]$axes
 
@@ -444,13 +444,13 @@ akde <- function(data,CTMM,VMM=NULL,debias=TRUE,weights=FALSE,smooth=TRUE,error=
       # calculate optimal bandwidth and some other information
       KDE[[i]] <- bandwidth(data=data[[i]],CTMM=CTMM[[i]],VMM=VMM[[i]],weights[i],verbose=TRUE,...)
     }
-    else if(class(CTMM)=="bandwidth") # bandwidth information was precalculated
+    else if(class(CTMM)[1]=="bandwidth") # bandwidth information was precalculated
     {
       KDE[[i]] <- CTMM[[i]]
       axes <- names(KDE[[i]]$h)
     }
     else
-    { stop(paste("CTMM argument is of class",class(CTMM))) }
+    { stop(paste("CTMM argument is of class",class(CTMM)[1])) }
 
     DEBIAS[[i]] <- ifelse(debias,KDE[[i]]$bias,FALSE)
   } # end loop over individuals
@@ -472,7 +472,8 @@ akde <- function(data,CTMM,VMM=NULL,debias=TRUE,weights=FALSE,smooth=TRUE,error=
   data.all <- do.call("rbind", data.all)
 
   # complete grid specification
-  grid <- kde.grid(data.all,H=H.all,axes=axes,alpha=error,res=res,dr=dr,grid=grid)
+  EXT <- extent(CTMM,level=1-error) # Gaussian extent (includes uncertainty)
+  grid <- kde.grid(data.all,H=H.all,axes=axes,alpha=error,res=res,dr=dr,grid=grid,EXT.min=EXT)
 
   # loop over individuals
   for(i in 1:n)
@@ -481,7 +482,7 @@ akde <- function(data,CTMM,VMM=NULL,debias=TRUE,weights=FALSE,smooth=TRUE,error=
 
     KDE[[i]] <- new.UD(KDE[[i]],info=attr(data[[i]],"info"),type='range',CTMM=ctmm())
     # in case bandwidth is pre-calculated...
-    if(class(CTMM0[[i]])=="ctmm") { attr(KDE[[i]],"CTMM") <- CTMM0[[i]] }
+    if(class(CTMM0[[i]])[1]=="ctmm") { attr(KDE[[i]],"CTMM") <- CTMM0[[i]] }
   }
 
   names(KDE) <- names(data)
@@ -520,7 +521,7 @@ prepare.H <- function(H,n,axes=c('x','y'))
 # construct a grid for the density function
 # non-destructuve WRT argument 'grid'
 # non-grid arguments are merely suggestions
-kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NULL,grid=NULL)
+kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NULL,EXT.min=NULL,grid=NULL)
 {
   H <- prepare.H(H,n=length(data$t),axes=axes) # (times,dim,dim)
 
@@ -530,11 +531,11 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   dH <- t(dH) # (times,dim)
 
   # format lazy grid arguments
-  if(!is.null(grid) && class(grid)=="list" && all(axes %in% names(grid))) # assuming coordinate list
+  if(!is.null(grid) && class(grid)[1]=="list" && all(axes %in% names(grid))) # assuming coordinate list
   { grid <- list(r=grid) }
-  else if(!is.null(grid) && class(grid) %nin% c("list","UD","RasterLayer"))
+  else if(!is.null(grid) && class(grid)[1] %nin% c("list","UD","RasterLayer"))
   {
-    if(class(grid)=="Extent")
+    if(class(grid)[1]=="Extent")
     { grid <- list(extent=grid) }
     else # assuming extent or dr
     {
@@ -550,17 +551,6 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   # recycle resolution if necessary
   if("dr" %in% names(grid)) { grid$dr <- array(grid$dr,length(axes)); names(grid$dr) <- axes }
 
-  # format extents canonically
-  get.extent <- function(EXT)
-  {
-    if(class(EXT)=="Extent")
-    { EXT <- cbind( c(EXT@xmin,EXT@xmax) , c(EXT@ymin,EXT@ymax) ) }
-    else # in case of data.frame
-    { EXT <- as.matrix(EXT) }
-
-    return(EXT)
-  }
-
   # centered sequence (mu) with n steps by by
   cseq <- function(mu,n,by)
   {
@@ -570,7 +560,7 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   }
 
   ### SPECIFY GRID ###
-  if(class(grid)=="RasterLayer") ### grid defined by raster ###
+  if(class(grid)[1]=="RasterLayer") ### grid defined by raster ###
   {
     EXT <- raster::extent(grid)
     EXT <- cbind( c(EXT@xmin,EXT@xmax) , c(EXT@ymin,EXT@ymax) )
@@ -597,7 +587,7 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   else if(!is.null(grid$dr) && !is.null(grid$extent)) ### grid fully pre-specified... with possible conflict ###
   { # resolution takes presidence over extent
     dr <- grid$dr
-    EXT <- get.extent(grid$extent)
+    EXT <- as.matrix(grid$extent)
     if(isTRUE(grid$align.to.origin)) { EXT <- t((round(t(EXT)/dr+1/2)-1/2)*dr) } # assume mostly correct
 
     dEXT <- EXT[2,]-EXT[1,]
@@ -614,7 +604,7 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   else if(!is.null(grid$extent)) ### grid extent specified, but not resolution ###
   {
     # align.to.origin doesn't make sense without dr specified
-    EXT <- get.extent(grid$extent)
+    EXT <- as.matrix(grid$extent)
     dEXT <- EXT[2,]-EXT[1,]
 
     if(is.null(dr)) { dr <- dEXT/res }
@@ -655,6 +645,8 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
     R <- get.telemetry(data,axes) # (times,dim)
 
     if(is.null(EXT)) { EXT <- rbind( apply(R-dH,2,min) , apply(R+dH,2,max) ) } # (ext,dim) }
+    colnames(EXT) <- axes
+    if(!is.null(EXT.min)) { EXT <- as.matrix(extent(list(EXT,EXT.min),level=1)) }
     dEXT <- EXT[2,]-EXT[1,]
 
     # grid center
@@ -769,7 +761,7 @@ kde <- function(data,H,axes=c("x","y"),bias=FALSE,W=NULL,alpha=0.001,res=NULL,dr
   { CDF <- pmf2cdf(PMF) }
 
   dV <- prod(dr)
-  result <- list(PDF=PMF/dV,CDF=CDF,r=R,dr=dr)
+  result <- list(PDF=PMF/dV,CDF=CDF,axes=axes,r=R,dr=dr)
   return(result)
 }