+++ NEWS.md
@@ -1,3 +1,22 @@
+ctmm 0.5.8 (2019-12-09)
+================
+
+  * improvements to ctmm.select stepwise selection, especially with error and/or circulation
+  
+  * improvements to ctmm.fit for nearly linear homeranges
+  
+  * bug fix in ctmm.loglike for BM/IOU models with error
+
+  * units of speed supported by %#% operator
+  
+  * new units argument in plot.outlie
+
+  * options(time.units='mean') and options(time.units='calendar') for %#% operator and display units
+  
+  * ctmm.select no longer warns when model features are not supported (ctmm.fit does)
+  
+  * compatibility fix for R version 4
+  
 ctmm 0.5.7 (2019-10-06)
 ================